requirements.txt
alpaca-py>=0.30.0
huggingface-hub>=0.25.0
groq>=0.12.0
gradio>=4.0.0
pandas>=2.0.0
APScheduler>=3.10.0
pytz>=2023.3
